+++ contracts/modules/ProtocolSettings.sol
@@ -12,8 +12,8 @@ import "../openzeppelin/SafeERC20.sol";
 import "../mixins/ProtocolTokenUser.sol";
 
 contract ProtocolSettings is State, ProtocolTokenUser, ProtocolSettingsEvents {
-    using SafeERC20 for IERC20;
-    using SafeMath for uint256;
+	using SafeERC20 for IERC20;
+	using SafeMath for uint256;
 
 	constructor() public {}
 
@@ -157,50 +157,38 @@ contract ProtocolSettings is State, ProtocolTokenUser, ProtocolSettingsEvents {
 		emit SetFeesController(msg.sender, oldController, newController);
 	}
 
-    function withdrawFees(
-        address token,
-        address receiver)
-        external
-        returns (uint)
-    {
-        require(msg.sender == feesController, "unauthorized");
+	function withdrawFees(address token, address receiver) external returns (uint256) {
+		require(msg.sender == feesController, "unauthorized");
 
-        uint lendingBalance = lendingFeeTokensHeld[token];
-        if (lendingBalance > 0) {
-            lendingFeeTokensHeld[token] = 0;
-            lendingFeeTokensPaid[token] = lendingFeeTokensPaid[token].add(lendingBalance);
-        }
+		uint256 lendingBalance = lendingFeeTokensHeld[token];
+		if (lendingBalance > 0) {
+			lendingFeeTokensHeld[token] = 0;
+			lendingFeeTokensPaid[token] = lendingFeeTokensPaid[token].add(lendingBalance);
+		}
 
-        uint tradingBalance = tradingFeeTokensHeld[token];
-        if (tradingBalance > 0) {
-            tradingFeeTokensHeld[token] = 0;
-            tradingFeeTokensPaid[token] = tradingFeeTokensPaid[token].add(tradingBalance);
-        }
+		uint256 tradingBalance = tradingFeeTokensHeld[token];
+		if (tradingBalance > 0) {
+			tradingFeeTokensHeld[token] = 0;
+			tradingFeeTokensPaid[token] = tradingFeeTokensPaid[token].add(tradingBalance);
+		}
 
-        uint borrowingBalance = borrowingFeeTokensHeld[token];
-        if (borrowingBalance > 0) {
-            borrowingFeeTokensHeld[token] = 0;
-            borrowingFeeTokensPaid[token] = borrowingFeeTokensPaid[token].add(borrowingBalance);
-        }
+		uint256 borrowingBalance = borrowingFeeTokensHeld[token];
+		if (borrowingBalance > 0) {
+			borrowingFeeTokensHeld[token] = 0;
+			borrowingFeeTokensPaid[token] = borrowingFeeTokensPaid[token].add(borrowingBalance);
+		}
 
-        uint amount = lendingBalance.add(tradingBalance).add(borrowingBalance);
-        if (amount == 0) {
-            return amount;
-        }
+		uint256 amount = lendingBalance.add(tradingBalance).add(borrowingBalance);
+		if (amount == 0) {
+			return amount;
+		}
 
-        IERC20(token).safeTransfer(receiver, amount);
+		IERC20(token).safeTransfer(receiver, amount);
 
-        emit WithdrawFees(
-            msg.sender,
-            token,
-            receiver,
-            lendingBalance,
-            tradingBalance,
-            borrowingBalance
-        );
+		emit WithdrawFees(msg.sender, token, receiver, lendingBalance, tradingBalance, borrowingBalance);
 
-        return amount;
-    }
+		return amount;
+	}
 
 	function withdrawLendingFees(
 		address token,
