+++ interfaces/ISovryn.sol
@@ -60,11 +60,7 @@ contract ISovryn is
 
 	function setFeesController(address newController) external;
 
-    function withdrawFees(
-        address token,
-        address receiver)
-        external
-        returns (uint);
+	function withdrawFees(address token, address receiver) external returns (uint256);
 
 	function withdrawLendingFees(
 		address token,
